quiet down some build warnings with Qt under OSX clang++

+++ win/Qt/qt_yndlg.cpp
@@ -184,7 +184,7 @@ char NetHackQtYnDialog::Exec()
 	show();
 	char choice=0;
 	char ch_esc=0;
-	for (uint i=0; i<ch.length(); i++) {
+	for (uint i=0; i< (uint) ch.length(); i++) {
 	    if (ch[i].unicode()=='q') ch_esc='q';
 	    else if (!ch_esc && ch[i].unicode()=='n') ch_esc='n';
 	}
